+++ docs/changelog.md
@@ -20,6 +20,15 @@ Requires libvips v8.15.2
   [#4147](https://github.com/lovell/sharp/pull/4147)
   [@ton11797](https://github.com/ton11797)
 
+* Expose PNG text chunks as `comments` metadata.
+  [#4157](https://github.com/lovell/sharp/pull/4157)
+  [@nkeynes](https://github.com/nkeynes)
+
+* Expose optional `precision` and `minAmplitude` parameters of `blur` operation.
+  [#4168](https://github.com/lovell/sharp/pull/4168)
+  [#4172](https://github.com/lovell/sharp/pull/4172)
+  [@marcosc90](https://github.com/marcosc90)
+
 ### v0.33.4 - 16th May 2024
 
 * Remove experimental status from `pipelineColourspace`.